Markdown parsed by default
Both shortname and longname now parse md by default.
This needed a new helper (contains_markdown), that allows us to check for markdown in an input. This is mainly used to check for markdown in the shortname, and trigger errors if true and no explicit key. It's also used to trigger an error if markdown + attempt to use plural without explicit plural term.
This has needed some updates to styles (mainly to fix footnotes, which had not been compatible with markdown).

Finally, options (parse_markdown_*) are still used, but set to default true. These can be removed altogether once feature added to support raw strings.